refactor: 채팅 메시지 디스플레이 유틸 함수 리팩토링
