--- src/entities/chat/utils/check-chat-room-id-exist.ts
@@ -1,9 +1,0 @@
-export const checkChatRoomIdExist = (roomId: number | undefined): roomId is number => {
-  if (roomId === undefined) {
-    console.error('roomId is undefined on ChatRoom component');
-
-    return false;
-  }
-
-  return true;
-};
