+++ src/entities/chat/utils/index.ts
@@ -1,7 +1,13 @@
 export { getTimeDiffText } from './get-time-diff-text';
 
-export { checkChatRoomIdExist } from './check-chat-room-id-exist';
-
 export { checkRoomStateExist } from './check-room-state-exist';
 
 export { compareDateAndSenderOfChat } from './compare-date-and-sender-of-chat';
+
+export { insertMessageGroupForDisplay } from './insert-message-group-for-display';
+
+export { handleConfirmTransactionImpl } from './handle-confirm-transaction-impl';
+
+export { openNotifyMessageDeleteModalImpl } from './open-notify-message-delete-modal-impl';
+
+export { handleDeleteMessageImpl } from './handle-delete-message-impl';
