+++ src/entities/chat/utils/insert-message-group-for-display/index.ts
@@ -1,0 +1,1 @@
+export { insertMessageGroupForDisplay } from './insert-message-group-for-display';
